+++ DigiDocs/Error.swift
@@ -4,8 +4,12 @@ enum CameraError: Error {
     case orientation
 }
 
-enum PDFError: Error {
+enum PDFError: LocalizedError {
     case context
     case noContent
     case framework(Error)
+
+    var errorDescription: String? {
+        return L10n.foobarAlertTitle
+    }
 }