+++ DigiDocs/SimulatorImagePickerController.swift
@@ -1,5 +1,7 @@
 import UIKit
 
+// Camera doesn't work in the simulator, so we use this when targeting the simulator
+
 #if targetEnvironment(simulator)
 final class SimulatorImagePickerController: UIViewController, ImagePickerControlling {
     var allowsEditing: Bool = false