+++ DigiDocsTests/AppTestEnvironment.swift
@@ -2,7 +2,7 @@
 import UIKit
 
 final class AppTestEnvironment {
-    var mainViewController: MainViewControlling
+    var homeViewController: HomeViewControlling
     var pdfService: PDFServicing
     var camera: Camerable
     var cameraOverlay: CameraOverlayViewControlling
@@ -11,7 +11,7 @@ final class AppTestEnvironment {
     private(set) var alertController: AlertControlling!
     private(set) var overlayAlertController: AlertControlling!
     private(set) var pdfAlertController: AlertControlling!
-    private(set) var mainController: MainControlling!
+    private(set) var homeController: HomeControlling!
     private(set) var cameraController: CameraControlling!
     private(set) var listController: ListControlling!
     private(set) var optionsController: OptionsControlling!
@@ -20,12 +20,13 @@ final class AppTestEnvironment {
     private(set) var shareController: ShareControlling!
     private(set) var appController: AppControlling!
 
-    init(mainViewController: MainViewControlling = MockMainViewController(),
+    init(homeViewController: HomeViewControlling = MockHomeViewController(),
          pdfService: PDFServicing = MockPDFService(),
          camera: Camerable = MockCamera(),
          cameraOverlay: CameraOverlayViewControlling = MockCameraOverlayViewController(),
          pdfViewController: PDFViewControlling = MockPDFViewController()) {
-        self.mainViewController = mainViewController
+        type(of: self).resetStaticMocks()
+        self.homeViewController = homeViewController
         self.pdfService = pdfService
         self.camera = camera
         self.cameraOverlay = cameraOverlay
@@ -50,28 +51,35 @@ final class AppTestEnvironment {
             try! fileManager.copyItem(at: origURL, to: pdfURL)
         }
     }
+
+    // MARK: - private
+
+    private static func resetStaticMocks() {
+        SimulatorImagePickerController.isSourceTypeAvailable = true
+    }
 }
 
 // MARK: - TestEnvironment
 
 extension AppTestEnvironment: TestEnvironment {
     func inject() {
-        alertController = AlertController(presenter: mainViewController)
-        mainController = MainController(viewController: mainViewController, pdfService: pdfService)
+        alertController = AlertController(presenter: homeViewController)
+        homeController = HomeController(viewController: homeViewController, pdfService: pdfService)
         overlayAlertController = AlertController(presenter: cameraOverlay)
         cameraController = CameraController(camera: camera, cameraOverlay: cameraOverlay,
                                             alertController: alertController,
                                             overlayAlertController: overlayAlertController)
-        listController = ListController(alertController: alertController, presenter: mainViewController,
+        listController = ListController(alertController: alertController, presenter: homeViewController,
                                         pdfService: pdfService)
-        optionsController = OptionsController(presenter: mainViewController)
+        // popoverView required for ipad, but safe to inject dummy view as we're only testing iphone
+        optionsController = OptionsController(presenter: homeViewController, popoverView: UIView())
         pdfAlertController = AlertController(presenter: pdfViewController)
         pdfController = PDFController(viewController: pdfViewController, alertController: alertController,
-                                      pdfAlertController: pdfAlertController, presenter: mainViewController,
+                                      pdfAlertController: pdfAlertController, presenter: homeViewController,
                                       pdfService: pdfService)
         namingController = NamingController(alertController: alertController, pdfService: pdfService)
-        shareController = ShareController(presenter: mainViewController)
-        appController = AppController(mainController: mainController, cameraController: cameraController,
+        shareController = ShareController(presenter: homeViewController)
+        appController = AppController(homeController: homeController, cameraController: cameraController,
                                       listController: listController, optionsController: optionsController,
                                       pdfController: pdfController, namingController: namingController,
                                       shareController: shareController, alertController: alertController)